+++ functions/index.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const MODULE_REQUIRE = 1
+	/* built-in */
+	
+	/* NPM */
+	, noda = require('noda')
+	
+	/* in-package */
+	;
+
+module.exports = noda.requireDir('.');